refactor: modularize card rendering
